+++ tda_polilineas.h
@@ -34,7 +34,7 @@ float polilinea_buscar_xmin(polilinea_t *p);
 
 float polilinea_buscar_ymax(polilinea_t *p);
 
-void polilinea_agrandar(polilinea_t *p, int a);
+void polilinea_agrandar(polilinea_t *p, float a);
 
 polilinea_t *leer_polilinea(FILE *f);
